Removed negative values for Expense entity model. Moved logic of the Value property to the BudgetChangeEntity. Updated tests to use this class instead Income/Expense.
