+++ src/BudgetBadgerWebApi.Domain/Entities/Expense.cs
@@ -4,18 +4,6 @@ namespace BudgetBadgerWebApi.Domain.Entities
 {
     public class Expense : BudgetChangeEntity
     {
-        public decimal Value
-        {
-            get => _value;
-            set
-            {
-                if (value >= 0m)
-                    throw new ArgumentException($"{nameof(Value)} of the {GetType().Name} can't be or bigger than or equal to 0");
-
-                _value = value;
-            }
-        }
-
         public List<HouseMemberExpense> HouseMemberExpenses { get; set; } = new();
     }
 }
