+++ src/BudgetBadgerWebApi.Domain/Entities/Income.cs
@@ -4,18 +4,6 @@ namespace BudgetBadgerWebApi.Domain.Entities
 {
     public class Income : BudgetChangeEntity
     {
-        public decimal Value
-        {
-            get => _value;
-            set
-            {
-                if (value <= 0m)
-                    throw new ArgumentException($"{nameof(Value)} of the {GetType().Name} can't be lower than or equal to 0");
-
-                _value = value;
-            }
-        }
-
         public int HouseMemberId { get; set; }
         public HouseMember HouseMember { get; set; } = new();
     }
