+++ test/Domain.Tests/TestData/Entities/BudgetChangeEntity/ValueNegativeDataAttribute.cs
@@ -1,15 +1,15 @@
 ﻿using System.Reflection;
 using Xunit.Sdk;
 
-namespace Domain.Tests.TestData.Entities.Expense
+namespace Domain.Tests.TestData.Entities.BudgetChangeEntity
 {
     internal class ValueNegativeDataAttribute : DataAttribute
     {
         public override IEnumerable<object[]> GetData(MethodInfo testMethod)
         {
             yield return new object[] { 0m };
-            yield return new object[] { 0.01m };
-            yield return new object[] { 509.99m };
+            yield return new object[] { -0.0001m };
+            yield return new object[] { -8569m };
         }
     }
 }
