--- test/Domain.Tests/TestData/Entities/Income/ValueNegativeDataAttribute.cs
@@ -1,15 +1,0 @@
-﻿using System.Reflection;
-using Xunit.Sdk;
-
-namespace Domain.Tests.TestData.Entities.Income
-{
-    internal class ValueNegativeDataAttribute : DataAttribute
-    {
-        public override IEnumerable<object[]> GetData(MethodInfo testMethod)
-        {
-            yield return new object[] { 0m };
-            yield return new object[] { -0.0001m };
-            yield return new object[] { -8569m };
-        }
-    }
-}
