+++ test/Domain.Tests/UnitTests/Entities/BudgetChangeEntityTests.cs
@@ -1,19 +1,19 @@
-﻿using BudgetBadgerWebApi.Domain.Entities;
-using Domain.Tests.TestData.Entities.Income;
+﻿using BudgetBadgerWebApi.Domain.Entities.Common;
+using Domain.Tests.TestData.Entities.BudgetChangeEntity;
 using FluentAssertions;
 
 namespace Domain.Tests.UnitTests.Entities
 {
     [Trait("Unit Tests", "Entities")]
-    public class IncomeTests
+    public class BudgetChangeEntityTests
     {
         [Theory]
         [ValueNegativeData]
         public void ValueProperty_Should_Throw_ArgumentException_For_Value_LowerThanOrEqualTo_Zero(decimal testValue)
         {
-            Income income = new();
+            BudgetChangeEntity budgetChangeEntity = new();
 
-            Action act = () => income.Value = testValue;
+            Action act = () => budgetChangeEntity.Value = testValue;
 
             act.Should().ThrowExactly<ArgumentException>();
         }
@@ -22,11 +22,11 @@ namespace Domain.Tests.UnitTests.Entities
         [ValuePositiveData]
         public void ValueProperty_Should_Set_WithoutException_For_Value_GreaterThan_0(decimal testValue)
         {
-            Income income = new();
+            BudgetChangeEntity budgetChangeEntity = new();
 
-            income.Value = testValue;
+            budgetChangeEntity.Value = testValue;
 
-            income.Value.Should().Be(testValue);
+            budgetChangeEntity.Value.Should().Be(testValue);
         }
     }
 }
